adds source code to pass 'CanGetBalanceFromTransactionHistoryTest'
