+++ Boolean.CSharp.Main/Bank/ITransaction.cs
@@ -8,9 +8,9 @@ namespace Boolean.CSharp.Main.Bank
 {
     public interface ITransaction
     {
-        decimal Balance { get; }
         decimal Amount { get; }
         DateTime Date { get; }
+        decimal Balance { get; }
         string TransactionType { get; }
     }
 }
