+++ Boolean.CSharp.Main/Bank/Transaction.cs
@@ -8,6 +8,7 @@ namespace Boolean.CSharp.Main.Bank
 {
     public class Transaction : ITransaction
     {
+
         public Transaction(decimal amount, DateTime date, decimal balance, string transactionType )
         {
             Amount = amount;
@@ -20,8 +21,9 @@ namespace Boolean.CSharp.Main.Bank
 
         public DateTime Date { get; }
 
-        public decimal Balance { get; set; }
-
+        public decimal Balance { get; }
         public string TransactionType { get; }
+
+       
     }
 }
