+++ Boolean.CSharp.Test/ExtensionTests.cs
@@ -1,4 +1,5 @@
 ﻿using Boolean.CSharp.Main;
+using Boolean.CSharp.Main.Bank.AccountTypes;
 using NUnit.Framework;
 using System;
 using System.Collections.Generic;
@@ -16,15 +17,20 @@ namespace Boolean.CSharp.Test
         {
             _extension = new Extension();
         }
-        [Test]
-        private void TestQuestion1()
-        {
 
-        }
         [Test]
-        private void TestQuestion2()
+        public void CanGetBalanceFromTransactionHistoryTest()
         {
+            decimal expectedBalance = 2500.00M;
+            CurrentAccount currentAccount = new CurrentAccount("Current");
+            currentAccount.MakeDeposit(1000.00M);
+            currentAccount.MakeDeposit(2000.00M);
+            currentAccount.MakeWithdrawal(500.00M);
 
+            decimal actualBalance = currentAccount.GetBalance();
+
+            Assert.That(actualBalance, Is.EqualTo(expectedBalance));
         }
+
     }
 }
